jvm specific bedrock simple executor
